Initial refactoring for SMT & Contination support in `trace_decoder` (#197)
* First rough plan on trait structure

* Now conditionally uses evm types at compile time

- Created type aliases for all `evm_arithmetization` types that we
  need to swap out depending on whether or not we are compiling for
  `mpt` or `smt`.

* Removed the default feature of "mpt"

- Realized that there is no clean way to disable a default feature when
  it absolutely needs to be disabled (eg. when "smt" is enabled).
- Also added a compile time check to enforce that one of these features
  is enabled.

* Renamed `processed_block_trace.rs` --> `processed_block_trace_mpt.rs`

- Prep for `mpt`/`smt` logic feature gating.

* SQUASH!!

Rough draft of SMT & continuation support

- Still need to remove feature gating stuff.

* Removed feature gating for MPT/SMT

- Realized that we can link again both versions of the library at once
  without conditional compilation.

* A bit more cleanup

* Hack to avoid jemalloc failure

* Requested changes for PR #197

* Fixed all clippy errors
