+++ trace_decoder/Cargo.toml
@@ -28,7 +28,9 @@ thiserror = { workspace = true }
 
 # Local dependencies
 mpt_trie = { version = "0.2.1", path = "../mpt_trie" }
-evm_arithmetization = { version = "0.1.3", path = "../evm_arithmetization" }
+
+evm_arithmetization_mpt = { package = "evm_arithmetization", version = "0.1.3", path = "../evm_arithmetization" }
+evm_arithmetization_smt = { package = "evm_arithmetization", version = "0.1.2", git = "https://github.com/0xPolygonZero/zk_evm.git", branch = "feat_type2_hack" }
 
 [dev-dependencies]
 pretty_env_logger = "0.5.0"
